refactor(ui): split history table content
Closes #2641
